+++ x-pack/apm-server/sampling/eventstorage/pebble.go
@@ -1,0 +1,7 @@
+package eventstorage
+
+import "github.com/cockroachdb/pebble"
+
+func OpenPebble(storageDir string) (*pebble.DB, error) {
+	return pebble.Open(storageDir, &pebble.Options{})
+}
